runtime:IOのloadでbyteではなくushortにキャストする

+++ runtime/Command.cs
@@ -146,7 +146,7 @@
                         case 2: //Load
                             data = onLoad();
                             for(int i = 0;i < 0x1000; i++)
-                                memory.Store((byte)(0xf000 + i), data[i]);
+                                memory.Store((ushort)(0xf000 + i), data[i]);
                             break;
                         case 3: //Save
                             onSave(data);
